+++ src/lib.rs
@@ -1,2 +1,3 @@
+pub mod correlator;
 pub mod observation;
 pub mod settings;